Split exchange and wallet balance refresh actions and timers
