+++ app/modules/exchanges.ts
@@ -5,7 +5,7 @@ import _ from 'lodash';
 import { Action, Dispatch, GetState, ThunkAction } from '../actions/actions.types';
 import { generateId } from '../helpers/reducers';
 import { getExchangeBalances } from '../pages/portfolio/selectors/selectBalances';
-import { getExchanges, getTimers, getWallets } from '../selectors/selectGlobalState';
+import { getExchanges } from '../selectors/selectGlobalState';
 import { unifySymbols } from '../utils/unifySymbols';
 import {
   AddExchangeAction,
@@ -26,11 +26,8 @@ import {
 } from './exchanges.types';
 import { GlobalState } from './index';
 import { requestTickerUpdate } from './ticker';
-import { setLastUpdate, startTimer } from './timer';
-import { fetchBalanceForWallet } from './wallets';
+import { EXCHANGE_BALANCE_TIMER, setLastUpdate } from './timer';
 
-const BALANCE_REFRESH_MS = 30000;
-// const TRADE_REFRESH_MS = 60000;
 
 // Reducer
 export default function reducer(state: Exchanges = {}, action: Action): Exchanges {
@@ -64,7 +61,7 @@ export function addExchange(type: string, credentials: ExchangeCredentials): Thu
       exchangeType: type,
       credentials,
     });
-    dispatch(fetchAllBalances());
+    dispatch(fetchAllExchangeBalances());
   };
 }
 
@@ -78,7 +75,7 @@ export function updateExchangeCredentials(
       id,
       credentials,
     });
-    dispatch(fetchAllBalances());
+    dispatch(fetchAllExchangeBalances());
   };
 }
 
@@ -97,15 +94,11 @@ export function deleteExchange(id: string): Action {
   };
 }
 
-// TODO Split into exchange and wallet balances
-export function fetchAllBalances(): ThunkAction {
+export function fetchAllExchangeBalances(): ThunkAction {
   return async (dispatch: Dispatch, getState: GetState) => {
-    dispatch(setLastUpdate('balances'));
+    dispatch(setLastUpdate(EXCHANGE_BALANCE_TIMER));
     const exchanges = getConfiguredExchanges(getState());
     await Promise.all(_.map(exchanges, exchange => dispatch(fetchBalancesForExchange(exchange))));
-
-    const wallets = getWallets(getState());
-    await Promise.all(_.map(wallets, wallet => dispatch(fetchBalanceForWallet(wallet))));
   };
 }
 
@@ -141,19 +134,6 @@ function failedRequest(id: string, error: string): FailedExchangeRequestAction {
 
 function getConfiguredExchanges(state: GlobalState): Exchanges {
   return state.exchanges;
-}
-
-export function continuouslyFetchBalances(): ThunkAction {
-  return (dispatch: Dispatch, getState: GetState) => {
-    if (!getTimers(getState()).balances) {
-      const balanceTimer = window.setInterval(
-        () => dispatch(fetchAllBalances()),
-        BALANCE_REFRESH_MS
-      );
-
-      dispatch(startTimer('balances', balanceTimer));
-    }
-  };
 }
 
 async function fetchBalances(exchange: Exchange) {
